Hosted Gallery - add first visit britain page, separate campaigns, and keep footer elements inline for mobile
