+++ commercial/app/controllers/commercial/HostedContentController.scala
@@ -9,7 +9,7 @@ import views.html.hosted.{guardianHostedGallery, guardianHostedVideo}
 class HostedContentController extends Controller {
 
   def renderHostedPage(campaignName: String, pageName: String) = Action { implicit request =>
-    HostedPage.fromPageName(pageName) match {
+    HostedPage.fromPageName(campaignName, pageName) match {
       case Some(page: HostedVideoPage) => Cached(60)(RevalidatableResult.Ok(guardianHostedVideo(page)))
       case Some(page: HostedGalleryPage) => Cached(60)(RevalidatableResult.Ok(guardianHostedGallery(page)))
       case _ => NoCache(NotFound)
