+++ common/app/common/commercial/HostedPage.scala
@@ -1,7 +1,5 @@
 package common.commercial
 
-import conf.Static
-import conf.switches.Switches
 import model.GuardianContentTypes.Hosted
 import model.{MetaData, SectionSummary, StandalonePage}
 import play.api.libs.json.{JsArray, JsNumber, JsString}
@@ -57,7 +55,7 @@ case class HostedVideoPage(
     )
   }
 
-  lazy val nextPage : HostedVideoPage = HostedPage.fromPageName(nextPageName).collect{case v: HostedVideoPage => v} getOrElse HostedPage.defaultPage
+  lazy val nextPage : HostedVideoPage = RenaultHostedPages.fromPageName(nextPageName).collect{case v: HostedVideoPage => v} getOrElse RenaultHostedPages.defaultPage
 }
 
 case class HostedGalleryPage(
@@ -123,126 +121,16 @@ case class HostedGalleryImage(
 
 object HostedPage {
 
-  private val teaserPageName = "design-competition-teaser"
-  private val episode1PageName = "design-competition-episode1"
-  private val episode2PageName = "design-competition-episode2"
-  private val galleryPageName = "gallery-test"
+  private val renaultCampaignName = "renault-car-of-the-future"
+  private val galleryTestCampaignName = "hosted-gallery"
+  private val visitBritainCampaignName = "visit-britain"
 
-  private val teaser: HostedVideoPage = HostedVideoPage(
-    pageUrl = "https://www.theguardian.com/commercial/advertiser-content/renault-car-of-the-future/design-competition-teaser",
-    pageName = teaserPageName,
-    standfirst = "Who better to dream up the cars of tomorrow than the people who'll be buying them? Students at Central St Martins are working with Renault to design the interior for cars that will drive themselves. Watch this short video to find out more about the project, and visit this page again soon to catch up on the students' progress",
-    logoUrl = Static("images/commercial/logo_renault.jpg"),
-    zoeUrl = Static("images/commercial/ren_zoe.jpg"),
-    bannerUrl = Static("images/commercial/ren_commercial_banner.jpg"),
-    video = HostedVideo(
-      mediaId = "renault-car-of-the-future",
-      title = "Designing the car of the future",
-      duration = 86,
-      posterUrl = Static("images/commercial/renault-video-poster.jpg"),
-      srcUrl = "https://multimedia.guardianapis.com/interactivevideos/video.php?file=160516GlabsTestSD"
-    ),
-    nextPageName = episode1PageName
-  )
-
-  private val episode1: HostedVideoPage = HostedVideoPage(
-    pageUrl = "https://www.theguardian.com/commercial/advertiser-content/renault-car-of-the-future/design-competition-episode1",
-    pageName = episode1PageName,
-    standfirst = "Renault challenged Central St Martins students to dream up the car of the future. The winning design will be announced at Clerkenwell Design Week (and on this site). Watch this short video to find out who made the shortlist",
-    logoUrl = Static("images/commercial/logo_renault.jpg"),
-    zoeUrl = Static("images/commercial/ren_zoe.jpg"),
-    bannerUrl = Static("images/commercial/ren_commercial_banner.jpg"),
-    video = HostedVideo(
-      mediaId = "renault-car-of-the-future",
-      title = "Renault shortlists 'car of the future' designs",
-      duration = 160,
-      posterUrl = Static("images/commercial/renault-video-poster-ep1.jpg"),
-      srcUrl = "https://multimedia.guardianapis.com/interactivevideos/video.php?file=160523GlabsRenaultTestHD"
-    ),
-    nextPageName = episode2PageName
-  )
-
-  private val episode2: HostedVideoPage = HostedVideoPage(
-    pageUrl = "https://www.theguardian.com/commercial/advertiser-content/renault-car-of-the-future/design-competition-episode2",
-    pageName = episode2PageName,
-    standfirst = "A group of Central St Martins students took part in a competition to dream up the car of the future. The winning design is radical and intriguing. Meet the team whose blue-sky thinking may have created a blueprint for tomorrow's autonomous cars",
-    logoUrl = Static("images/commercial/logo_renault.jpg"),
-    zoeUrl = Static("images/commercial/ren_zoe.jpg"),
-    bannerUrl = Static("images/commercial/ren_commercial_banner.jpg"),
-    video = HostedVideo(
-      mediaId = "renault-car-of-the-future",
-      title = "Is this the car of the future?",
-      duration = 158,
-      posterUrl = Static("images/commercial/renault-video-poster-ep2.jpg"),
-      srcUrl = "http://multimedia.guardianapis.com/interactivevideos/video.php?file=160603GlabsRenaultTest3"
-    ),
-    nextPageName = episode1PageName
-  )
-
-  private val omgbImages : List[HostedGalleryImage] = List(
-    HostedGalleryImage (
-      url = "http://static.theguardian.com/commercial/hosted/gallery-prototype/omgb1.jpg",
-      title = "Finding the giants of Heligan",
-      caption = "The Lost Gardens of Heligan, Cornwall, England. These gardens, along with their giants, were lost for 25 years under a tangle of weeds before being restored to take their place among the finest gardens in Great Britain."
-    ),
-    HostedGalleryImage (
-      url = "http://static.theguardian.com/commercial/hosted/gallery-prototype/omgb2.jpg",
-      title = "Rolling hills and rugged moorland",
-      caption = "Malham Tarn Estate, Yorkshire Dales, England. Get your boots on and start exploring this popular Yorkshire Dales beauty spot."
-    ),
-    HostedGalleryImage (
-      url = "http://static.theguardian.com/commercial/hosted/gallery-prototype/omgb3.jpg",
-      title = "A walker’s paradise",
-      caption = "Isle of Skye, Scotland. From family walks on the spectacular coastline to dramatic scrambles in the Cuillin mountains, the island provides fantastic walking for everyone. Photo by Lars Scheider"
-    ),
-    HostedGalleryImage (
-      url = "http://static.theguardian.com/commercial/hosted/gallery-prototype/omgb4.jpg",
-      title = "Walking on water",
-      caption = "Loch Lomond & The Trossachs National Park, Scotland. The National Park stretches from the incredible mountains and glens of the Trossachs to the vast tranquil beauty of Loch Lomond. Photo by Bestjobers"
-    ),
-    HostedGalleryImage (
-      url = "http://static.theguardian.com/commercial/hosted/gallery-prototype/omgb5.jpg",
-      title = "Running across the plains",
-      caption = "Isle of Harris, Scotland. The Isle of Harris isn’t actually an island, it’s the southern and more mountainous part of Lewis and Harris, the largest island in the Outer Hebrides. Photo by Lars Scheider"
-    ),
-    HostedGalleryImage (
-      url = "http://static.theguardian.com/commercial/hosted/gallery-prototype/omgb6.jpg",
-      title = "Relaxing at Loch Maree",
-      caption = "Loch Maree, Scotland. More than 60 islands dot Loch Maree and the area is perfect for bird watching and walking."
-    ),
-    HostedGalleryImage (
-      url = "http://static.theguardian.com/commercial/hosted/gallery-prototype/omgb7.jpg",
-      title = "The ruins of Tintern Abbey",
-      caption = "Monmouthshire, Wales. The ruins of this 12th-century abbey are famous for inspiring the works of notable English artists, including Romantic poet William Wordsworth and painter Thomas Gainsborough."
-    ),
-    HostedGalleryImage (
-      url = "http://static.theguardian.com/commercial/hosted/gallery-prototype/omgb8.png",
-      title = "Ride the Hogwarts Express!",
-      caption = "Glenfinnan Viaduct, Scotland. The Jacobite steam train is a great way to experience the stunning scenery and special atmosphere of Glenfinnan. You may even recognise it from Harry Potter! Photo by Colin Roberts."
-    )
-  )
-
-  private val gallery: HostedGalleryPage = HostedGalleryPage(
-    images = omgbImages,
-    pageUrl = "https://www.theguardian.com/commercial/advertiser-content/hosted-gallery/gallery-test",
-    pageName = galleryPageName,
-    title = "Great Britain - Home of Amazing Moments #OMGB",
-    ctaText = "Explore our collection of unique experiences from all over Great Britain.",
-    ctaLink = "http://en.omgb.com/map/",
-    ctaIndex = 5,
-    standfirst = "Welcome to Great Britain, a country to be explored, experienced and discovered. See for yourself and discover the moments you'll want to share.",
-    logoUrl = "http://static.theguardian.com/commercial/hosted/gallery-prototype/omgb.png"
-  )
-
-  lazy val defaultPage = teaser
-
-  def fromPageName(pageName: String): Option[HostedPage] = {
-    pageName match {
-      case `teaserPageName` => Some(teaser)
-      case `episode1PageName` => Some(episode1)
-      case `episode2PageName` => Some(episode2)
-      case `galleryPageName` if Switches.hostedGalleryTest.isSwitchedOn => Some(gallery)
-      case _ => None
+  def fromPageName(campaignName: String, pageName: String): Option[HostedPage] = {
+    campaignName match {
+      case `renaultCampaignName` => RenaultHostedPages.fromPageName(pageName)
+      case `galleryTestCampaignName` => HostedGalleryTestPage.fromPageName(pageName)
+      case `visitBritainCampaignName` => VisitBritainHostedPages.fromPageName(pageName)
+      case _ => None;
     }
   }
 }
