+++ static/src/javascripts/projects/common/modules/commercial/hosted-gallery.js
@@ -5,23 +5,10 @@ define([
     'fastdom',
     'common/utils/$',
     'qwery',
-    'common/utils/ajax',
     'common/utils/config',
     'common/utils/detect',
     'common/utils/fsm',
     'common/utils/mediator',
-    'common/utils/template',
-    'common/utils/url',
-    'common/modules/component',
-    'common/modules/ui/blockSharing',
-    'common/modules/ui/images',
-    'common/views/svgs',
-    'text!common/views/content/block-sharing.html',
-    'text!common/views/content/button.html',
-    'text!common/views/content/endslate.html',
-    'text!common/views/content/loader.html',
-    'text!common/views/content/share-button.html',
-    'text!common/views/content/share-button-mobile.html',
     'lodash/collections/map',
     'lodash/functions/throttle',
     'lodash/collections/forEach',
@@ -34,23 +21,10 @@ define([
     fastdom,
     $,
     qwery,
-    ajax,
     config,
     detect,
     FiniteStateMachine,
     mediator,
-    template,
-    url,
-    Component,
-    blockSharing,
-    imagesModule,
-    svgs,
-    blockSharingTpl,
-    buttonTpl,
-    endslateTpl,
-    loaderTpl,
-    shareButtonTpl,
-    shareButtonMobileTpl,
     map,
     throttle,
     forEach,
